requirements.txt
alembic==1.13.1
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
diskcache==5.6.3
filelock==3.15.4
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
greenlet==3.0.3
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
optuna==3.5.0
packaging==23.2
pandas==2.2.1
petastorm==0.12.1
pillow==10.2.0
playsound==1.3.0
psutil==5.9.8
py4j==0.10.9.7
pyarrow==15.0.0
pygame==2.6.1
pyparsing==3.1.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.28
sympy==1.13.1
threadpoolctl==3.3.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
